Move daemontools commands to separate module
